Added all the part 3 of Libft and updated MakeFile

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: ltomas-d <ltomas-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/29 19:36:12 by ltomas-d          #+#    #+#             */
-/*   Updated: 2026/04/29 20:11:37 by ltomas-d         ###   ########.fr       */
+/*   Updated: 2026/05/04 12:36:19 by ltomas-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*ptr;
-	int		len_s;
+	size_t	len_s;
 
 	if (s == NULL)
 		return (NULL);
